# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Picroom API
  version: 1.0.0
  description: |
    Picroom — self-hosted image hosting service.

    Two API surfaces:
    1. REST/JSON under `/api/v1/` — first-class management API.
    2. AWS S3-compatible under `/s3/` — drop-in replacement for object storage
       clients (PicGo, rclone, AWS CLI).

    All endpoints require authentication except where noted.
  license:
    name: MIT
    identifier: MIT
  contact:
    name: Picroom Maintainers
    url: https://github.com/picroom/picroom
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://{host}
    description: Self-hosted production
    variables:
      host:
        default: picroom.example.com
...
paths:
...
  /api/v1/teams:
    get:
      tags: [teams]
      summary: List teams current user belongs to
      operationId: listMyTeams
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
    post:
      tags: [teams]
      summary: Create a team
      operationId: createTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeamRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/v1/teams/{teamId}:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    get:
      tags: [teams]
      summary: Get a team
      operationId: getTeam
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/teams/{teamId}/members:
    parameters:
      - $ref: '#/components/parameters/TeamId'
    get:
      tags: [teams]
      summary: List team members
      operationId: listTeamMembers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'
    post:
      tags: [teams]
      summary: Add a member to a team
      operationId: addTeamMember
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: picroom_session
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sigv4:
      type: apiKey
      in: header
      name: Authorization
      description: AWS Signature Version 4
  parameters:
...
    TeamId:
      in: path
      name: teamId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          examples: [quota_exceeded, validation_failed]
        message:
          type: string
          description: Human-readable message
        request_id:
          type: string
          format: uuid
        details:
          type: object
          additionalProperties: true
...
    User:
      type: object
      required: [id, email, name, role, created_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum: [viewer, uploader, manager, admin]
        avatar_url:
          type: string
          format: uri
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Team:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
          nullable: true
        storage_policy:
          type: string
          default: default
        created_at:
          type: string
          format: date-time
    CreateTeamRequest:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
          pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
        description:
          type: string
    TeamMember:
      type: object
      required: [user, role, joined_at]
      properties:
        user:
          $ref: '#/components/schemas/User'
        role:
          type: string
          enum: [viewer, uploader, manager, admin]
        joined_at:
          type: string
          format: date-time
    AddMemberRequest:
      type: object
      required: [user_id, role]
      properties:
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [viewer, uploader, manager, admin]
...
security:
  - cookieAuth: []
  - bearerAuth: []